+++ wsgidav/server/server_cli.py
@@ -436,8 +436,8 @@ def _run_cheroot(app, config, _server):
         _logger.error("Try `pip install cheroot`.")
         return False
 
-    version = wsgi.Server.version
-    version = f"WsgiDAV/{__version__} {version} Python {util.PYTHON_VERSION}"
+    org_version = wsgi.Server.version
+    version = f"WsgiDAV/{__version__} {org_version} Python {util.PYTHON_VERSION}"
     wsgi.Server.version = version
 
     info = _get_common_info(config)