+++ CHANGELOG.md
@@ -2,12 +2,21 @@
 
 ## 4.3.0 / Unreleased
 
+- BREAKING:
+  - The new option `fs_dav_provider.follow_symlinks` is false by default, 
+    since [symlinks may be a security risk](https://serverfault.com/questions/244592/followsymlinks-on-apache-why-is-it-a-security-risk).
+  - The new option `fs_dav_provider.shadow_map` replaces the (undocumented)
+    argument `FilesystemProvider(..., shadow={...})`
 - Install pam_dc dependencies using extra syntax: `pip install wsgidav[pam]`
 - #281 Requesting range off end of file does not return 416 status code
 - #282 Hotfix PUT request without content-length (fix for Finder on MacOS Ventura)
 - Add `logging.enable` option to activate the 'wsgidav' logger when this package
   is used as a library. This replaces an explicit call to `utils.init_logging()`.
   When running as CLI, this option is on by default.
+- Add `fs_dav_provider` section to options.
+- Add `fs_dav_provider.follow_symlinks` option to follow symlinks 
+- Add `fs_dav_provider.shadow_map`
+- Remove (unused) interface definitions
 
 ## 4.2.0 / 2023-02-18
 
@@ -33,7 +42,7 @@
 
 ## 4.0.2 / 2022-08-01
 
-- #245: Add charset utf-8 to Content-Type header
+- #245 Add charset utf-8 to Content-Type header
 - #246 Add dir_browser/htdocs folder to sdist (.tar.gz)
 - #248 Provider does not support set_last_modified" error
 - #251 Fix removing locks in recursive mode