+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - BREAKING:
   - See option `fs_dav_provider.follow_symlinks` below.
-- Stop testing Python 3.7 (EOL: 2023-06-27)
+  - Stop testing Python 3.7 (EOL: 2023-06-27)
 - Install pam_dc dependencies using extra syntax: `pip install wsgidav[pam]`
 - #281 Requesting range off end of file does not return 416 status code
 - #282 Hotfix PUT request without content-length (fix for Finder on MacOS Ventura)
@@ -22,6 +22,7 @@
 - Add `fs_dav_provider.shadow_map`, which can be used to blend in a favicon
   when context is opened inline (#230)
 - Remove (unused) interface definitions
+- Use the [Furo](https://github.com/pradyunsg/furo) documentation theme
 
 ## 4.2.0 / 2023-02-18